+++ methods/process.py
@@ -1,16 +1,23 @@
-"""
-In this .py file we lay out the pipelines for the various methods of
-    1. Tokenization
-    2. Vectorization
-    3. Regression/model
-"""
+"""Main file for the experiment/process through pipelines"""
+# Importing the pipeline
 from sklearn.pipeline import Pipeline
 
+# Vectorizers
+from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
+
+# Classifiers
 from sklearn.svm import SVC
 from sklearn.naive_bayes import MultinomialNB
 from sklearn.linear_model import LogisticRegression
 from sklearn.ensemble import RandomForestClassifier
-from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
+
+# Neural imports
+from methods.neural import NeuralNetwork
+
+
+# To delete later
+import pandas as pd
+from sklearn.model_selection import train_test_split
 
 
 # For passing onto all random states:
@@ -64,17 +71,47 @@ def build_pipeline(pipeline_model: str) -> Pipeline:
                 random_state=__RANDOM_SEED__
             )
         },
-
+        'cnn': {
+            'neural': NeuralNetwork(
+                model_type='cnn'
+            )
+        },
+        'lstm': {
+            'neural': NeuralNetwork(
+                model_type='lstm',
+                # epochs=1,  # TODO: DELETE THESE!!
+                # early_stop=True  # ALSO THIS
+            )
+        }
     }
+
+    # Select the model using .get -> eval to True/False
     selected_model = models.get(pipeline_model)
     if selected_model:
-        return Pipeline([
-            ('vectorizer', selected_model['vectorizer']),
-            ('classifier', selected_model['classifier'])
-        ])
+        if 'vectorizer' in selected_model and 'classifier' in selected_model:
+            return Pipeline([
+                ('vectorizer', selected_model['vectorizer']),
+                ('classifier', selected_model['classifier'])
+            ])
+        elif 'neural' in selected_model:
+            return Pipeline([
+                ('neural', selected_model['neural'])])
+
+        else:
+            raise KeyError(f"Invalid key: {pipeline_model}. Choose from {list(models.keys())}.")
     else:
         raise KeyError(f"Invalid key: {pipeline_model}. Choose from {list(models.keys())}.")
 
 
+class Experiment:
+    pass
+
+
 if __name__ == '__main__':
-    pipeline = build_pipeline('logisic')
+    pipeline = build_pipeline('lstm')
+    df = pd.read_csv('../data/cleaned_extrovert.csv', engine='pyarrow')
+    x_train, x_test, y_train, y_test = train_test_split(df['post'], df['label'],
+                                                        test_size=0.2, random_state=__RANDOM_SEED__)
+    pipeline.fit(x_train, y_train)
+    y_pred = pipeline.predict(x_test)
+